Add error handling middleware.

+++ server/src/services/TodoService.ts
@@ -1,4 +1,5 @@
 import {TodoModel} from "../models/Todo";
+import {ApiError} from "../exceptions/ApiError";
 
 
 class TodoService {
@@ -7,14 +8,25 @@ class TodoService {
     }
 
     async create(text: string) {
+        if (!text) throw ApiError.BadRequest('Text is required');
         return TodoModel.create({text});
     }
 
     async update(id: string, text: string, done: boolean) {
-        return TodoModel.findByIdAndUpdate(id, {text, done});
+        if(!id) throw ApiError.BadRequest('Id is required');
+
+        const todo = await TodoModel.findById(id);
+        if(!todo) throw ApiError.NotFound();
+
+        return TodoModel.findByIdAndUpdate(id, {text, done}, {returnOriginal: false});
     }
 
     async delete(id: string) {
+        if(!id) throw ApiError.BadRequest('Id is required');
+
+        const todo = await TodoModel.findById(id);
+        if(!todo) throw ApiError.NotFound();
+
         return TodoModel.findByIdAndDelete(id);
     }
